+++ .eslintrc.json
@@ -9,6 +9,8 @@
         "@typescript-eslint"
     ],
     "rules": {
+		"no-unused-vars": "warn",
+		"no-unused-expressions": "warn",
         "@typescript-eslint/naming-convention": "warn",
         "@typescript-eslint/semi": "warn",
         "curly": "warn",
